+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- New color for log running files
+- Console reporter small updates
+
+### Changed
+- Workflow `pr_master.yml` file. 
+
 ## 0.0.2 - 2024-11-20
 ### Added
 - `src` for runner